+++ source/component/position.d
@@ -9,11 +9,11 @@ final class Position : Component
   mixin TypeDecl;
   
   vec2 position;
-  float angle;
+  double angle;
   
   alias position this;
   
-  this(vec2 position, float angle)
+  this(vec2 position, double angle)
   {
     this.position = position;
     this.angle = angle;
